qdev: realize machine container
Long term we want to create machines as unrealized objects, and then
realize everything outside the constructor functions.  This patch
prepares for this by adding a place where the machine is realized.
As long as calls to qdev_init and qdev_init_nofail exist, the
propagation of realize will correctly stop there.

Based on a patch from Andreas Faerber.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>