target-i386: remove setting tsc-frequency from x86_def_t
Setting tsc-frequency from x86_def_t is NOP because default tsc_khz
in x86_def_t is 0 and CPUX86State.tsc_khz is also initialized to 0
by default. So there is not need to set ovewrite tsc_khz with default
0 because field was already initialized to 0.

custom tsc-frequency setting is not affected due to it is being set
without using x86_def_t (previous patch)

Field tsc_khz in x86_def_t becomes unused with this patch, so drop it
as well.

Signed-off-by: Igor Mammedov <imammedo@redhat.com>
Reviewed-by: Eduardo Habkost <ehabkost@redhat.com>
---
 v2:
  - rebased with "target-i386: move out CPU features initialization
    in separate func" patch dropped