virtio: add auxiliary sg items for ramblock straddling
Virtio requests that straddle ramblocks cause the assertion "virtio: Trying to
map MMIO memory". These requests have to be split in two requests and be mapped
separately to two different host memory ranges (one in each dimm/ramblock).

This has been observed with stress testing on VMs with hotplugged DIMMs.

With this patch, virtio_map_sg will create the auxiliary sg elements and will
insert each auliary sg after the originating sg item. The following sg items
are shifted accordingly.

Note: simply appending sg items at the end of the list, does not work because
the virtio code often assumes the last sg is a header sg. e.g. the blk_inhdr for
virtio-blk devices. We tried appending at the end of the list, and manually
shifting/swapping the original header item to the end of the list for these
cases. But even then there were some rare I/O problems when an auxuliary sg
item is not placed next to the originating sg item. The solution of splitting
the original sg in two at the original place in the sg list appears to work for
all cases.